pyproject.toml
[project]
name = "daypilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.173.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "mcp-weather-server>=0.1.3",
    "mcp[cli]>=1.9.4",
    "nest-asyncio>=1.6.0",
    "notion-client>=2.4.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
